[FIX] Premature exit of build process; Fixes #32
Although not quite certain why the issue arises, my assumption is that
because the configuration file has to be named 'webextension-toolbox.js'
and the main toolbox script, located in bin/, has the same name,
somewhere along the execution the wrong file is exectued, when using
windows command lines like cmd or powershell, leading to an abrupt exit.

Previous assumption that the 'find-up' module, used for accessing the
configuration file, is the cause of the error seem to be false, as build
process stops before even requiring the module.